# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: AMUPF Digital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (President/Executive)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: is_approved
          in: query
          schema:
            type: boolean
        - name: department_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/approve:
    post:
      operationId: approveKyc
      tags: [users]
      summary: Approve KYC for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/reject:
    post:
      operationId: rejectKyc
      tags: [users]
      summary: Reject KYC for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectKycInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required: [user_id, full_name, email, role, student_id, is_approved, is_active, created_at]
      properties:
        user_id:
          type: string
        full_name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, member, executive, president, higher_official]
        student_id:
          type: string
        phone:
          type: string
        department_id:
          type: ["integer", "null"]
        department_name:
          type: ["string", "null"]
        profile_photo:
          type: ["string", "null"]
        id_front_url:
          type: ["string", "null"]
        id_back_url:
          type: ["string", "null"]
        is_approved:
          type: boolean
        is_active:
          type: boolean
        rejection_reason:
          type: ["string", "null"]
        created_at:
          type: string
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        department_id:
          type: integer
        profile_photo:
          type: string
        role:
          type: string
        is_active:
          type: boolean
    RejectKycInput:
      type: object
      required: [rejection_reason]
      properties:
        rejection_reason:
          type: string
          minLength: 20

    # --- DEPARTMENT ---
...